ecom_backend/requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.2.0
boto3==1.34.108
botocore==1.34.108
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.7
cssselect2==0.7.0
dj-database-url==2.2.0
Django==4.2.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-csv-export-view==2.0.0
django-environ==0.10.0
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
firebase-admin==6.5.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
html5lib==1.1
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.2
msgpack==1.1.0
oscrypto==1.3.0
Pillow==10.1.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.1.4
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.1.0
python-dateutil==2.9.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.6
reportlab==4.0.9
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
